pyproject.toml
[tool.poetry]
name = "tg-crypto-bot"
version = "0.1.0"
description = "Crypto tg bot"
authors = ["Maksim <maxikpetlitskii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.1.1"
requests = "^2.31.0"
pandas = "^2.2.2"
plotly = "^5.22.0"
kaleido = "0.1.0post1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
